Enhance error handling and exception management in mod teardown and main execution flow
- Updated the `tear_down` method in `ModHandler` to collect exceptions and raise an `ExceptionGroup` if multiple errors occur during module teardown.
- Modified the `run` function in `main.py` to re-raise exceptions after handling, ensuring that unexpected errors are not silently caught.
- Improved type handling for pandas Series in `mod.py` to explicitly set the data type, enhancing data integrity.
- Introduced an `integration_test` function in the testing module to facilitate integration testing with result validation.

+++ rqalpha/mod/rqalpha_mod_sys_analyser/mod.py
@@ -470,8 +470,8 @@ class AnalyserMod(AbstractMod):
             result_dict["summary"]["excess_max_drawdown_duration_end_date"] = str(max_ddd.end_date)
             result_dict["summary"]["excess_max_drawdown_duration_days"] = (max_ddd.end_date - max_ddd.start_date).days
         else:
-            weekly_b_returns = pandas.Series(index=weekly_returns.index)
-            monthly_b_returns = pandas.Series(index=monthly_returns.index)
+            weekly_b_returns = pandas.Series(index=weekly_returns.index, dtype=float)
+            monthly_b_returns = pandas.Series(index=monthly_returns.index, dtype=float)
 
         # 周度风险指标
         weekly_risk = Risk(weekly_returns, weekly_b_returns, risk_free_rate, WEEKLY)
